+++ BazMap.lua
@@ -424,9 +424,4 @@ addon.config.onLoad = function(self)
 
     BazCore:RegisterOptionsTable(ADDON_NAME .. "-Settings", GetSettingsPage)
     BazCore:AddToSettings(ADDON_NAME .. "-Settings", "Settings", ADDON_NAME)
-
-    BazCore:RegisterOptionsTable(ADDON_NAME .. "-Profiles", function()
-        return BazCore:GetProfileOptionsTable(ADDON_NAME)
-    end)
-    BazCore:AddToSettings(ADDON_NAME .. "-Profiles", "Profiles", ADDON_NAME)
 end